Hash#merge! should have the specs and #update be the alias
